move non-command function to a sep file
